Move `bits` iterator
